hook up quick jQuery-free modal
bootstrap has all of the styles we need (already included). trying to do this without jQuery if we don't need it, for file size reasons.

One hack: I put the background (modalBackdrop) inside mainModal, so that it doesn't capture click events when it's invisible and /does/ when it's present. Just seemed like a more natural DOM structure.
But that means I need to override its z-index in bootstrap so that it's not in front of the modal contents. In designs, it's not clear that there's a backdrop at all, so we might just want to delete that backdrop altogether.

TODO:
* populate the rest of the modal (images, captions, etc)
* hook up right-left arrows
* hook up close arrow
* populate the state data (ugh)
* somehow remap clicks on the svg states to real US states (that data doesn't seem to be present in the us.json blob)