Disable full-width video on large screens
Aims to address "Can the video view port be a perfect square with a white background so there's more space above/below the video?"

This makes videos 75% wide rather than 100% wide on screens over 770px, so that more of the content below the video is visible withouthaving to scroll.

Admittedly, I don't know I am correctly interpreting the request as written above.